Resurrect support for daemonization
Going to great lengths to make frame state consistent regardless of
whether emacs is initialized with a display or detached (i.e., when
daemonized). See the init-display function for a summary of the
differences.

I also had to remove the use-package harness around all core config
setup because it hijacked the expected control flow for making a frame
during debugging.